+++ core/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- `addToLabelIndex`: a function to include a thing in a label index, so that it can be found via search
+
 ## 0.13.0
 
 ### Changed